refactor: Update project structure
